upd: hover notifications to pause/pin
