+++ frontend/src/app/app.component.scss
@@ -79,129 +79,6 @@
   }
 }
 
-.toast-container {
-  position: fixed;
-  top: 5rem;
-  right: 3rem;
-  z-index: 999;
-  max-width: 25%;
-  // padding: 1rem;
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-
-  @media (width < 765px) {
-    max-width: 80%;
-    right: 1rem;
-    padding: 0.5rem;
-  }
-}
-
-.toast {
-  position: relative;
-  overflow: hidden;
-  padding: 0.75rem 1rem;
-  border-radius: 0.5rem;
-  background-color: var(--color-surface-container-high);
-  color: var(--color-on-surface);
-  animation: slideInRight 0.3s ease-out forwards;
-  cursor: pointer;
-}
-
-.toast-header {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.toast-icon {
-  width: 1.25rem;
-  height: 1.25rem;
-  flex-shrink: 0;
-  fill: currentColor;
-}
-
-.toast-message {
-  flex: 1;
-  text-wrap: wrap;
-}
-
-.toast-close,
-.toast-pin {
-  background: none;
-  border: none;
-  cursor: pointer;
-  padding: 0;
-  color: inherit;
-  opacity: 0.7;
-  flex-shrink: 0;
-
-  svg {
-    width: 1rem;
-    height: 1rem;
-    fill: currentColor;
-    display: block;
-  }
-
-  &:hover {
-    opacity: 1;
-  }
-}
-
-.toast-pin[aria-pressed='true'] {
-  opacity: 1;
-}
-
-.toast-progress {
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  height: 3px;
-  width: 100%;
-  background: rgba(255, 255, 255, 0.4);
-  animation: toastProgress 3s linear forwards;
-}
-
-@keyframes toastProgress {
-  from {
-    width: 100%;
-  }
-
-  to {
-    width: 0%;
-  }
-}
-
-.toast.pinned .toast-progress {
-  display: none;
-}
-
-.info {
-  background-color: var(--color-info);
-  color: var(--color-info-text);
-}
-
-.success {
-  background-color: var(--color-success);
-  color: var(--color-success-text);
-}
-
-.error {
-  background-color: var(--color-warning);
-  color: var(--color-warning-text);
-}
-
-@keyframes slideInRight {
-  from {
-    transform: translateX(100%);
-    opacity: 0;
-  }
-
-  to {
-    transform: translateX(0);
-    opacity: 1;
-  }
-}
 
 .dialog-content {
   padding: 1rem;
