+++ frontend/src/app/app.component.ts
@@ -1,18 +1,18 @@
 import {AsyncPipe} from '@angular/common';
 import {ChangeDetectionStrategy, Component, ElementRef, HostListener, inject, OnDestroy, OnInit, signal, viewChild} from '@angular/core';
 import {Router, RouterOutlet} from '@angular/router';
-import {Subscription} from 'rxjs';
 import {msMinute} from 'src/util';
 import {TimeRemainingPipe} from './helpers/time-remaining.pipe';
 import {SidenavComponent} from './nav/sidenav/sidenav.component';
 import {TopnavComponent} from './nav/topnav/topnav.component';
+import {NotificationsComponent} from './notifications/notifications.component';
 import {AuthService} from './services/auth.service';
-import {MessageData, WebsocketService} from './services/websocket.service';
+import {WebsocketService} from './services/websocket.service';
 import {LoadIndicatorComponent} from './shared/load-indicator';
 
 @Component({
   selector: 'app-root',
-  imports: [AsyncPipe, LoadIndicatorComponent, RouterOutlet, TimeRemainingPipe, TopnavComponent, SidenavComponent],
+  imports: [AsyncPipe, LoadIndicatorComponent, NotificationsComponent, RouterOutlet, TimeRemainingPipe, TopnavComponent, SidenavComponent],
   templateUrl: './app.component.html',
   styleUrl: './app.component.scss',
   changeDetection: ChangeDetectionStrategy.OnPush,
@@ -22,11 +22,6 @@ export class AppComponent implements OnDestroy, OnInit {
   private readonly router = inject(Router);
   private readonly websocketService = inject(WebsocketService);
 
-  protected messages = signal<MessageData[]>([]);
-  protected readonly pinnedToasts = signal(new Set<MessageData>());
-  private readonly toastTimeouts = new Map<MessageData, ReturnType<typeof setTimeout>>();
-  private toastSubscription?: Subscription;
-
   private extendTimeoutId: any;
   private sessionTimeoutId: any;
   private readonly IDLE_LIMIT: number = 12 * msMinute;
@@ -34,17 +29,7 @@ export class AppComponent implements OnDestroy, OnInit {
   protected sessionEndTime = signal<number>(Date.now() + this.IDLE_LIMIT + this.EXTEND_LIMIT);
 
   ngOnInit() {
-    // Reset the idle timer
     this.resetIdleTimer();
-
-    // Subscribe to messages and display them
-    this.toastSubscription = this.websocketService.toastMessage.subscribe({
-      next: (data: MessageData) => {
-        this.messages.update((msgs) => [...msgs, data]);
-        const id = setTimeout(() => this.removeMessage(data), 3000);
-        this.toastTimeouts.set(data, id);
-      },
-    });
   }
 
   // Uncomment the below code to enable mouse movement detection too!
@@ -63,9 +48,7 @@ export class AppComponent implements OnDestroy, OnInit {
   }
 
   closeAllSubscriptions() {
-    this.showEndedDialog();
     this.websocketService.close();
-    this.toastSubscription?.unsubscribe();
     this.authService.logout().subscribe(() => {
       this.router.navigate(['/login']);
     });
@@ -79,29 +62,10 @@ export class AppComponent implements OnDestroy, OnInit {
     if (this.sessionTimeoutId) {
       clearTimeout(this.sessionTimeoutId);
     }
-    this.toastTimeouts.forEach((id) => clearTimeout(id));
-  }
-
-  removeMessage(data: MessageData): void {
-    clearTimeout(this.toastTimeouts.get(data));
-    this.toastTimeouts.delete(data);
-    this.pinnedToasts.update((s) => {
-      const n = new Set(s);
-      n.delete(data);
-      return n;
-    });
-    this.messages.update((msgs) => msgs.filter((m) => m !== data));
-  }
-
-  stopAutoClose(data: MessageData): void {
-    clearTimeout(this.toastTimeouts.get(data));
-    this.toastTimeouts.delete(data);
-    this.pinnedToasts.update((s) => new Set(s).add(data));
   }
 
   // Non-required: dialogs are only in the DOM when isAuthenticated() is true
   readonly sessionEndingDialog = viewChild<ElementRef<HTMLDialogElement>>('sessionEndingDialog');
-  readonly sessionEndedDialog = viewChild<ElementRef<HTMLDialogElement>>('sessionEndedDialog');
 
   showEndingDialog(): void {
     this.sessionEndTime.set(Date.now() + this.EXTEND_LIMIT);
@@ -109,11 +73,6 @@ export class AppComponent implements OnDestroy, OnInit {
     this.sessionTimeoutId = setTimeout(() => {
       this.closeAllSubscriptions();
     }, this.EXTEND_LIMIT);
-  }
-
-  showEndedDialog(): void {
-    this.closeEndingDialog();
-    this.sessionEndedDialog()?.nativeElement.showModal();
   }
 
   closeEndingDialog(): void {
@@ -126,7 +85,4 @@ export class AppComponent implements OnDestroy, OnInit {
     this.closeEndingDialog();
   }
 
-  reloadPage(): void {
-    this.router.navigate(['/login']);
-  }
 }
